Worked on tests

+++ tests/fdatetime_test_fat_date_time.c
@@ -632,7 +632,7 @@ int fdatetime_test_fat_date_time_get_string_size(
 	FDATETIME_TEST_ASSERT_EQUAL_INT(
 	 "string_size",
 	 string_size,
-	 22 );
+	 (size_t) 22 );
 
         FDATETIME_TEST_ASSERT_IS_NULL(
          "error",
@@ -659,7 +659,7 @@ int fdatetime_test_fat_date_time_get_string_size(
 	FDATETIME_TEST_ASSERT_EQUAL_INT(
 	 "string_size",
 	 string_size,
-	 16 );
+	 (size_t) 16 );
 
         FDATETIME_TEST_ASSERT_IS_NULL(
          "error",
